+++ crates/e2m2e-propagation/src/lib.rs
@@ -5,6 +5,7 @@
 pub mod abm;
 pub mod butcher;
 pub mod cowell;
+pub mod lambert;
 pub mod multistep_methods;
 pub mod pd45;
 pub mod pd78;