+++ e2m2e/integrators.py
@@ -6,7 +6,14 @@ import numpy as np
 import numpy.typing as npt
 
 try:
-    from e2m2e._integrators import CowellResult, MultistepMethod, MultistepResult, RkMethod
+    from e2m2e._integrators import (
+        CowellResult,
+        MultistepMethod,
+        MultistepResult,
+        RkMethod,
+        lambert_batch_py,
+        lambert_izzo_py,
+    )
     from e2m2e._integrators import cowell_step as _cowell_step
     from e2m2e._integrators import multistep_step as _multistep_step
     from e2m2e._integrators import rk_step as _rk_step
@@ -17,6 +24,8 @@ except ModuleNotFoundError:
     MultistepMethod = None  # type: ignore[misc,assignment]
     MultistepResult = None  # type: ignore[misc,assignment]
     RkMethod = None  # type: ignore[misc,assignment]
+    lambert_batch_py = None  # type: ignore[misc,assignment]
+    lambert_izzo_py = None  # type: ignore[misc,assignment]
     _cowell_step = None  # type: ignore[misc,assignment]
     _multistep_step = None  # type: ignore[misc,assignment]
     _rk_step = None  # type: ignore[misc,assignment]
@@ -31,6 +40,8 @@ __all__ = [
     "cowell_step",
     "CowellResult",
     "initialize_cowell_history",
+    "lambert_izzo_py",
+    "lambert_batch_py",
 ]